+++ src/main/java/si/matjazcerkvenik/eventlogger/parsers/FluentdTailParser.java
@@ -37,24 +37,27 @@ public class FluentdTailParser implements IEventParser {
 
             GsonBuilder builder = new GsonBuilder();
             Gson gson = builder.create();
-            List<DEvent> elist = gson.fromJson(dRequest.getBody().trim(), new TypeToken<List<DEvent>>(){}.getType());
+            List<DEvent> eventList = gson.fromJson(dRequest.getBody().trim(), new TypeToken<List<DEvent>>(){}.getType());
 
-            String ident = "null";
             long now = System.currentTimeMillis();
 
-            for (DEvent de : elist) {
-                de.setId(DProps.increaseAndGetEventsReceivedCount());
+            for (DEvent de : eventList) {
                 de.setRuntimeId(DProps.RUNTIME_ID);
                 de.setTimestamp(now);
                 de.setHost(dRequest.getRemoteHost());
                 de.setEventSource(dRequest.getRemoteHost());
-                de.setLogfile("unknown");
                 de.setEndpoint(dRequest.getRequestUri());
-                de.setIdent("http.post");
-                de.setPid("0");
                 if (dRequest.getParameterMap().containsKey("ident")) {
-                    ident = dRequest.getParameterMap().get("ident");
-                    de.setIdent(ident);
+                    de.setIdent(dRequest.getParameterMap().get("ident"));
+                } else {
+                    DMetrics.eventlogger_events_ignored_total.labels(dRequest.getRemoteHost(), dRequest.getRequestUri(), "missing ident").inc();
+                    eventList.remove(de);
+                    continue;
+                }
+                if (de.getMessage() == null || de.getMessage().trim().length() == 0) {
+                    DMetrics.eventlogger_events_ignored_total.labels(dRequest.getRemoteHost(), dRequest.getRequestUri(), "no content").inc();
+                    eventList.remove(de);
+                    continue;
                 }
                 if (dRequest.getParameterMap().containsKey("pid")) {
                     de.setPid(dRequest.getParameterMap().get("pid"));
@@ -65,21 +68,20 @@ public class FluentdTailParser implements IEventParser {
                 if (dRequest.getParameterMap().containsKey("tag")) {
                     de.setTag(dRequest.getParameterMap().get("tag"));
                 }
+                de.setId(DProps.increaseAndGetEventsReceivedCount());
                 LogFactory.getLogger().trace(de.toString());
-//			System.out.println(de.toString());
+                DMetrics.eventlogger_events_total.labels(dRequest.getRemoteHost(), de.getHost(), de.getIdent()).inc();
             }
 
-            if (elist != null && elist.size() > 0) {
-                DMetrics.eventlogger_events_total.labels(dRequest.getRemoteHost(), dRequest.getRemoteHost(), ident).inc(elist.size());
-                return elist;
+            if (eventList != null && eventList.size() > 0) {
+                return eventList;
             }
 
-            DMetrics.eventlogger_events_ignored_total.labels(dRequest.getRemoteHost(), dRequest.getMethod()).inc();
-            LogFactory.getLogger().warn("HttpWebhook: doPost: message is empty; event will be ignored");
+            LogFactory.getLogger().warn("FluentdTailParser: eventList is empty!");
             return null;
 
         } catch (Exception e) {
-            LogFactory.getLogger().error("FluentdTailParser: parseRequest: Exception: " + e.getMessage());
+            LogFactory.getLogger().error("FluentdTailParser: Exception: " + e.getMessage());
             throw new EventParserException("fluentd-tail parser failed");
         }